♻️ extract rum batch logic + introduce RUM_EVENT_COLLECTED
